MinutesX/requirements.txt
# MinutesX: Intelligent Multi-Agent Meeting Notes
# Requirements

# =============================================================================
# Google ADK & AI
# =============================================================================
google-adk>=0.1.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.38.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# =============================================================================
# Google Meet & Calendar Integration
# =============================================================================
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1

# =============================================================================
# Web Framework
# =============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# =============================================================================
# Async & Concurrency
# =============================================================================
aiohttp>=3.9.0
asyncio>=3.4.3
httpx>=0.25.0

# =============================================================================
# A2A Protocol & Message Bus
# =============================================================================
redis>=5.0.0
aioredis>=2.0.0

# =============================================================================
# Memory & Vector Store
# =============================================================================
faiss-cpu>=1.7.4
chromadb>=0.4.0
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# =============================================================================
# NLP & Evaluation
# =============================================================================
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.0

# =============================================================================
# Observability
# =============================================================================
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
structlog>=23.2.0

# =============================================================================
# Integrations
# =============================================================================
slack-sdk>=3.23.0
jira>=3.5.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.0
rich>=13.7.0
tenacity>=8.2.0

# =============================================================================
# Testing
# =============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# =============================================================================
# Development
# =============================================================================
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
